backup_pythonanywhere/requirements_complete.txt
# 🏥 Roteiro de Dispensação - COMPLETO para PythonAnywhere
# Todas as dependências para projeto full-stack

# === CORE FLASK ===
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# === AI & ML (Sistema RAG Completo) ===
openai==1.12.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0

# === PROCESSAMENTO DE TEXTO AVANÇADO ===
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
regex==2023.10.3

# === BASE DE DADOS E CACHE ===
redis==5.0.1
# sqlite3 - Já incluído no Python, não precisa instalar

# === UTILIDADES CORE ===
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
typing-extensions==4.8.0

# === SEGURANÇA E VALIDAÇÃO ===
bleach==6.0.0
validators==0.22.0

# === MONITORAMENTO E LOGS ===
python-logging==1.0.0

# === PERFORMANCE ===
joblib==1.3.2
tqdm==4.66.1

# === COMPATIBILIDADE ===
six==1.16.0
certifi==2023.7.22
urllib3==2.0.7
charset-normalizer==3.3.2
idna==3.4

# === PROCESSAMENTO JSON/YAML ===
PyYAML==6.0.1

# === DATETIME ===
python-dateutil==2.8.2

# === MARKDOWN PROCESSING ===
markdown==3.5.1
markupsafe==2.1.3